demo app changes and useStore test
