+++ demo/src/routeTree.gen.ts
@@ -12,8 +12,8 @@
 
 import { Route as rootRoute } from "./routes/__root";
 import { Route as TodoImport } from "./routes/todo";
+import { Route as SharedImport } from "./routes/shared";
 import { Route as PersistentImport } from "./routes/persistent";
-import { Route as GlobalImport } from "./routes/global";
 import { Route as CounterImport } from "./routes/counter";
 import { Route as ContextImport } from "./routes/context";
 import { Route as AsyncImport } from "./routes/async";
@@ -27,15 +27,15 @@ const TodoRoute = TodoImport.update({
   getParentRoute: () => rootRoute,
 } as any);
 
-const PersistentRoute = PersistentImport.update({
-  id: "/persistent",
-  path: "/persistent",
+const SharedRoute = SharedImport.update({
+  id: "/shared",
+  path: "/shared",
   getParentRoute: () => rootRoute,
 } as any);
 
-const GlobalRoute = GlobalImport.update({
-  id: "/global",
-  path: "/global",
+const PersistentRoute = PersistentImport.update({
+  id: "/persistent",
+  path: "/persistent",
   getParentRoute: () => rootRoute,
 } as any);
 
@@ -95,18 +95,18 @@ declare module "@tanstack/react-router" {
       preLoaderRoute: typeof CounterImport;
       parentRoute: typeof rootRoute;
     };
-    "/global": {
-      id: "/global";
-      path: "/global";
-      fullPath: "/global";
-      preLoaderRoute: typeof GlobalImport;
-      parentRoute: typeof rootRoute;
-    };
     "/persistent": {
       id: "/persistent";
       path: "/persistent";
       fullPath: "/persistent";
       preLoaderRoute: typeof PersistentImport;
+      parentRoute: typeof rootRoute;
+    };
+    "/shared": {
+      id: "/shared";
+      path: "/shared";
+      fullPath: "/shared";
+      preLoaderRoute: typeof SharedImport;
       parentRoute: typeof rootRoute;
     };
     "/todo": {
@@ -126,8 +126,8 @@ export interface FileRoutesByFullPath {
   "/async": typeof AsyncRoute;
   "/context": typeof ContextRoute;
   "/counter": typeof CounterRoute;
-  "/global": typeof GlobalRoute;
   "/persistent": typeof PersistentRoute;
+  "/shared": typeof SharedRoute;
   "/todo": typeof TodoRoute;
 }
 
@@ -136,8 +136,8 @@ export interface FileRoutesByTo {
   "/async": typeof AsyncRoute;
   "/context": typeof ContextRoute;
   "/counter": typeof CounterRoute;
-  "/global": typeof GlobalRoute;
   "/persistent": typeof PersistentRoute;
+  "/shared": typeof SharedRoute;
   "/todo": typeof TodoRoute;
 }
 
@@ -147,8 +147,8 @@ export interface FileRoutesById {
   "/async": typeof AsyncRoute;
   "/context": typeof ContextRoute;
   "/counter": typeof CounterRoute;
-  "/global": typeof GlobalRoute;
   "/persistent": typeof PersistentRoute;
+  "/shared": typeof SharedRoute;
   "/todo": typeof TodoRoute;
 }
 
@@ -159,8 +159,8 @@ export interface FileRouteTypes {
     | "/async"
     | "/context"
     | "/counter"
-    | "/global"
     | "/persistent"
+    | "/shared"
     | "/todo";
   fileRoutesByTo: FileRoutesByTo;
   to:
@@ -168,8 +168,8 @@ export interface FileRouteTypes {
     | "/async"
     | "/context"
     | "/counter"
-    | "/global"
     | "/persistent"
+    | "/shared"
     | "/todo";
   id:
     | "__root__"
@@ -177,8 +177,8 @@ export interface FileRouteTypes {
     | "/async"
     | "/context"
     | "/counter"
-    | "/global"
     | "/persistent"
+    | "/shared"
     | "/todo";
   fileRoutesById: FileRoutesById;
 }
@@ -188,8 +188,8 @@ export interface RootRouteChildren {
   AsyncRoute: typeof AsyncRoute;
   ContextRoute: typeof ContextRoute;
   CounterRoute: typeof CounterRoute;
-  GlobalRoute: typeof GlobalRoute;
   PersistentRoute: typeof PersistentRoute;
+  SharedRoute: typeof SharedRoute;
   TodoRoute: typeof TodoRoute;
 }
 
@@ -198,8 +198,8 @@ const rootRouteChildren: RootRouteChildren = {
   AsyncRoute: AsyncRoute,
   ContextRoute: ContextRoute,
   CounterRoute: CounterRoute,
-  GlobalRoute: GlobalRoute,
   PersistentRoute: PersistentRoute,
+  SharedRoute: SharedRoute,
   TodoRoute: TodoRoute,
 };
 
@@ -217,8 +217,8 @@ export const routeTree = rootRoute
         "/async",
         "/context",
         "/counter",
-        "/global",
         "/persistent",
+        "/shared",
         "/todo"
       ]
     },
@@ -234,11 +234,11 @@ export const routeTree = rootRoute
     "/counter": {
       "filePath": "counter.tsx"
     },
-    "/global": {
-      "filePath": "global.tsx"
-    },
     "/persistent": {
       "filePath": "persistent.tsx"
+    },
+    "/shared": {
+      "filePath": "shared.tsx"
     },
     "/todo": {
       "filePath": "todo.tsx"
