+++ demo/src/routes/todo.tsx
@@ -5,6 +5,39 @@ export const Route = createFileRoute("/todo")({
   component: RouteComponent,
 });
 
+interface TodoState {
+  input: string;
+  todos: { title: string; complete: boolean }[];
+}
+
+const store = createStore(
+  {
+    input: "",
+    todos: [],
+  } as TodoState,
+  (set, get) => ({
+    setInput: (input: string) => set({ input }),
+    addTodo: () => {
+      if (!get().input) {
+        return;
+      }
+      set((state) => ({
+        todos: [...state.todos, { title: state.input, complete: false }],
+        input: "",
+      }));
+    },
+    toggleTodo: (index: number) =>
+      set((state) => ({
+        todos: state.todos.map((todo, i) => {
+          if (index === i) {
+            return { ...todo, complete: !todo.complete };
+          }
+          return todo;
+        }),
+      })),
+  }),
+);
+
 function RouteComponent() {
   const {
     state: { input, todos },
@@ -28,48 +61,21 @@ function RouteComponent() {
         />
         <button type="submit">Add</button>
       </form>
-      <ul>
-        {todos.map((todo, index) => (
-          <li
-            key={todo.title}
-            data-testid={`todo-${index}`}
-            role="button"
-            onClick={() => toggleTodo(index)}
-            className={`list-disc list-inside ${todo.complete && "line-through"}`}
-          >
-            {todo.title}
-          </li>
-        ))}
-      </ul>
+      {todos.length > 0 && (
+        <ul>
+          {todos.map((todo, index) => (
+            <li
+              key={todo.title}
+              data-testid={`todo-${index}`}
+              role="button"
+              onClick={() => toggleTodo(index)}
+              className={`list-disc list-inside ${todo.complete && "line-through"}`}
+            >
+              {todo.title}
+            </li>
+          ))}
+        </ul>
+      )}
     </div>
   );
 }
-
-interface TodoState {
-  input: string;
-  todos: { title: string; complete: boolean }[];
-}
-
-const store = createStore(
-  {
-    input: "",
-    todos: [],
-  } as TodoState,
-  (set) => ({
-    setInput: (input: string) => set({ input }),
-    addTodo: () =>
-      set((state) => ({
-        todos: [...state.todos, { title: state.input, complete: false }],
-        input: "",
-      })),
-    toggleTodo: (index: number) =>
-      set((state) => ({
-        todos: state.todos.map((todo, i) => {
-          if (index === i) {
-            return { ...todo, complete: !todo.complete };
-          }
-          return todo;
-        }),
-      })),
-  }),
-);
